lift Computed Values up
