+++ tic-tac-toe/src/components/GameBoard.jsx
@@ -1,20 +1,7 @@
-const initialGameBoard = [
-	[null, null, null],
-	[null, null, null],
-	[null, null, null],
-];
-
-export default function GameBoard({ onSelectSquare, gameTurn }) {
-	let gameBoard = initialGameBoard;
-
-	gameTurn?.forEach((turn) => {
-		const { rowIndex, cellIndex } = turn?.square;
-		gameBoard[rowIndex][cellIndex] = turn?.player;
-	});
-
+export default function GameBoard({ onSelectSquare, board }) {
 	return (
 		<ol id="game-board">
-			{gameBoard.map((row, rowIndex) => (
+			{board.map((row, rowIndex) => (
 				<li key={rowIndex}>
 					<ol>
 						{row.map((cell, cellIndex) => (
